Fix some fields in Domain classes, move Fixture to Domain, add refs in summary, fix project dependencies versions
