+++ AirCompany.Domain/Models/AircraftFamily.cs
@@ -1,27 +1,27 @@
 ﻿namespace AirCompany.Domain.Models;
 
 /// <summary>
-/// Represents an aircraft family (e.g., A320 Family, Boeing 737 Family).
+/// Represents an <see cref="AircraftFamily"/> (e.g., "A320 Family", "Boeing 737 Family").
 /// </summary>
 public class AircraftFamily
 {
     /// <summary>
-    /// Unique identifier of the aircraft family.
+    /// Unique identifier of the <see cref="AircraftFamily"/>.
     /// </summary>
-    public int Id { get; set; }
+    public required int Id { get; set; }
 
     /// <summary>
-    /// Name of the aircraft family (e.g., "A320 Family").
+    /// Name of the <see cref="AircraftFamily"/> (e.g., "A320 Family").
     /// </summary>
     public required string Name { get; set; }
 
     /// <summary>
-    /// Manufacturer of the aircraft family (e.g., "Airbus").
+    /// Manufacturer of this <see cref="AircraftFamily"/> (e.g., "Airbus").
     /// </summary>
     public required string Manufacturer { get; set; }
 
     /// <summary>
-    /// Navigation property: list of aircraft models in this family.
+    /// Collection of <see cref="AircraftModel"/>s belonging to this <see cref="AircraftFamily"/>.
     /// </summary>
-    public List<AircraftModel> Models { get; set; } = [];
+    public List<AircraftModel>? Models { get; set; } = [];
 }
