+++ AirCompany.Domain/Models/AircraftModel.cs
@@ -1,42 +1,42 @@
 ﻿namespace AirCompany.Domain.Models;
 
 /// <summary>
-/// Represents a specific aircraft model (e.g., A320neo).
+/// Represents a specific <see cref="AircraftModel"/> (e.g., "A320neo").
 /// </summary>
 public class AircraftModel
 {
     /// <summary>
-    /// Unique identifier of the aircraft model.
+    /// Unique identifier of the <see cref="AircraftModel"/>.
     /// </summary>
-    public int Id { get; set; }
+    public required int Id { get; set; }
 
     /// <summary>
-    /// Name of the aircraft model (e.g., "A320neo").
+    /// Name of the <see cref="AircraftModel"/> (e.g., "A320neo").
     /// </summary>
     public required string ModelName { get; set; }
 
     /// <summary>
-    /// Navigation property: the family this model belongs to.
+    /// The <see cref="AircraftFamily"/> this <see cref="AircraftModel"/> belongs to.
     /// </summary>
     public required AircraftFamily Family { get; set; }
 
     /// <summary>
-    /// Maximum flight range of the aircraft in kilometers.
+    /// Maximum flight range of this <see cref="AircraftModel"/> in kilometers.
     /// </summary>
     public required double FlightRangeKm { get; set; }
 
     /// <summary>
-    /// Maximum passenger capacity of the aircraft.
+    /// Maximum passenger capacity of this <see cref="AircraftModel"/>.
     /// </summary>
     public required int PassengerCapacity { get; set; }
 
     /// <summary>
-    /// Maximum cargo capacity of the aircraft in kilograms.
+    /// Maximum cargo capacity of this <see cref="AircraftModel"/> in kilograms.
     /// </summary>
-    public required decimal CargoCapacityKg { get; set; }
+    public required double CargoCapacityKg { get; set; }
 
     /// <summary>
-    /// Navigation property: list of flights using this aircraft model.
+    /// Collection of <see cref="Flight"/>s operated by this <see cref="AircraftModel"/>.
     /// </summary>
-    public List<Flight> Flights { get; set; } = [];
+    public List<Flight>? Flights { get; set; } = [];
 }
