+++ AirCompany.Domain/Models/Flight.cs
@@ -1,14 +1,15 @@
 ﻿namespace AirCompany.Domain.Models;
 
+
 /// <summary>
-/// Represents a scheduled flight.
+/// Represents a scheduled <see cref="Flight"/> between two airports.
 /// </summary>
 public class Flight
 {
     /// <summary>
-    /// Unique identifier of the flight.
+    /// Unique identifier of the <see cref="Flight"/>.
     /// </summary>
-    public int Id { get; set; }
+    public required int Id { get; set; }
 
     /// <summary>
     /// Flight code (e.g., "SU1234").
@@ -16,37 +17,37 @@ public class Flight
     public required string Code { get; set; }
 
     /// <summary>
-    /// Departure airport (could be IATA code or city name).
+    /// Departure airport for this <see cref="Flight"/> (IATA code or city name).
     /// </summary>
     public required string DepartureAirport { get; set; }
 
     /// <summary>
-    /// Arrival airport (could be IATA code or city name).
+    /// Arrival airport for this <see cref="Flight"/> (IATA code or city name).
     /// </summary>
     public required string ArrivalAirport { get; set; }
 
     /// <summary>
-    /// Date and time of departure.
+    /// Scheduled departure date and time of this <see cref="Flight"/>.
     /// </summary>
     public DateTime? DepartureDateTime { get; set; }
 
     /// <summary>
-    /// Date and time of arrival.
+    /// Scheduled arrival date and time of this <see cref="Flight"/>.
     /// </summary>
     public DateTime? ArrivalDateTime { get; set; }
 
     /// <summary>
-    /// Total flight duration.
+    /// Total duration of this <see cref="Flight"/>.
     /// </summary>
     public TimeSpan? FlightDuration { get; set; }
 
     /// <summary>
-    /// Navigation property: aircraft model for this flight.
+    /// The <see cref="AircraftModel"/> operating this <see cref="Flight"/>.
     /// </summary>
     public required AircraftModel AircraftModel { get; set; }
 
     /// <summary>
-    /// Navigation property: tickets issued for this flight.
+    /// Collection of <see cref="Ticket"/>s issued for this <see cref="Flight"/>.
     /// </summary>
-    public List<Ticket> Tickets { get; set; } = [];
+    public List<Ticket>? Tickets { get; set; } = [];
 }
